refactor(compare-environments): split into folder, fix review findings
Address PR review feedback on the Compare Environments redesign:

- Move the single 607-line component into a CompareEnvironments/ folder
  with a barrel: types.ts, useEnvironmentComparison hook (fetch + diff
  derivation, now unit-testable), ExpandedRow, CompareFeatureRow, and the
  layout component.
- Replace relative imports with components/ and common/ aliases.
- Reuse the shared FlagsmithValue type instead of an inline union.
- Surface a fetch error state with a retry action instead of silently
  swallowing failures and rendering "No differences found".
- Guard against stale overlapping comparisons with a request-id check so a
  slow earlier response can't overwrite a newer one.
- Make the row expand toggle keyboard accessible (role, tabIndex,
  aria-expanded, Enter/Space handling).
- Pass the numeric projectId into CreateFlagModal.
- Fall back to the left project flag for the right-side edit action so the
  modal edits the existing feature instead of switching to create mode.
- Consolidate duplicated diff header styles into a shared base rule.
