+++ frontend/web/components/CompareEnvironments/index.ts
@@ -1,0 +1,2 @@
+export { default } from './CompareEnvironments'
+export type { FeatureChange } from './types'
