+++ frontend/web/styles/3rdParty/_react-diff.scss
@@ -97,22 +97,22 @@
 }
 
 // Environment name legend for diff headers
-.diff-removed-header {
-  background-color: $danger-alfa-16;
-  color: $danger;
+.diff-removed-header,
+.diff-added-header {
   font-weight: 500;
   font-size: $font-caption;
   padding: 4px 8px;
   border-radius: 4px;
 }
 
+.diff-removed-header {
+  background-color: $danger-alfa-16;
+  color: $danger;
+}
+
 .diff-added-header {
   background-color: $success-alfa-16;
   color: $success;
-  font-weight: 500;
-  font-size: $font-caption;
-  padding: 4px 8px;
-  border-radius: 4px;
 }
 
 .diff-sign {
